requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0  
drf-spectacular==0.26.3
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-ses==4.4.0